Update r2j_pg_support.cpp
We have had issues with MetaCollectionsResult returning incorrect data, such as a total count of 100 records when more than 5000 were expected.

To address this, I have removed the reliance on this metric for tracking records remaining. 

I have also created a schema and table-set for pricelists, the reason for this is that pricelists have the children pricelistsItems. So we are now tracking all organization pricelists and using that to populate the ID field for pricelistsItems via pricelist_id.